requirements.txt
kombu==4.6.8
pocsuite3==1.5.9
secrets==1.0.2
records==0.5.3
psutil==5.6.7
tqdm==4.41.0
shodan==1.21.3
loguru==0.4.0
Flask==1.1.1
cloudscraper==1.2.16
urllib3==1.25.3
Flask_RESTful==0.3.8
aiohttp==3.6.2
tldextract==2.2.2
celery==4.4.2
Flask_Cors==3.0.8
simplejson==3.17.0
fire==0.2.1
aiomultiprocess==0.7.0
requests==2.22.0
tablib==0.14.0
Flask_PyMongo==2.3.0
cdx_toolkit==0.9.25
exrex==0.10.5
pymongo==3.10.1
python_nmap==0.6.1
beautifulsoup4==4.9.0
dnspython==1.16.0
tenacity
Flask-HTTPAuth
python-nmap
lxml